chore: rename `jwt` package
