+++ pkg/jwt/context/jwt.go
@@ -1,14 +1,14 @@
-package jwt
+package context
 
 import (
 	"context"
 	"fmt"
 
-	"github.com/ribeirohugo/go_middlewares/pkg/authentication"
+	jwtAuth "github.com/ribeirohugo/go_middlewares/pkg/jwt"
 )
 
 // GetClaims allows to extract claims from context.
-func (j *JWT) GetClaims(ctx context.Context) (authentication.Claims, error) {
+func (j *JWT) GetClaims(ctx context.Context) (jwtAuth.Claims, error) {
 	return j.auth.ParseClaims(ctx)
 }
