+++ pkg/jwt/context/middleware.go
@@ -1,4 +1,4 @@
-package jwt
+package context
 
 import (
 	"context"
